% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cesty.R
\name{cesty}
\alias{cesty}
\title{Roads (cesty)}
\format{\code{sf} data frame with 430 rows of 3 variables + geometry:

\describe{
  \item{oddelovac}{Median type of road.}
  \item{trasa}{Route (primary or unknown).}
  \item{typ}{Type of road}
}}
\source{
\url{http://www.diva-gis.org/gdata}
}
\usage{
cesty()
}
\description{
Function returning roads of Slovakia as \code{sf} polygon.
}
\details{
Due to stage of package development, this data is downloaded with the package.
}
