% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hranice.R
\name{hranice}
\alias{hranice}
\title{Borders (hranice)}
\format{
\code{sf} data frame with 1 row of geometry:
}
\source{
\url{http://www.diva-gis.org/gdata}
}
\usage{
hranice()
}
\description{
Function returning boundaries of Slovakia as \code{sf} polygon.
}
\details{
Due to stage of package development, this data is downloaded with the package.
}
