% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jazera.R
\name{jazera}
\alias{jazera}
\title{Lakes (jazera)}
\format{\code{sf} data frame with 13 rows of 1 variable + geometry:

\describe{
  \item{meno}{Name of the lake or inland water (jazera).}
}}
\source{
\url{http://www.diva-gis.org/gdata}
}
\usage{
jazera()
}
\description{
Function returning data frame of lakes and inland water of Slovakia as \code{sf} polygons.
}
\details{
Due to stage of package development, this data is downloaded with the package.
}
