% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kraje.R
\name{kraje}
\alias{kraje}
\title{Regions (kraje)}
\format{\code{sf} data frame with 8 rows of 1 variable + geometry:

\describe{
  \item{kraj}{Name of the region (kraj).}
}}
\source{
\url{http://www.diva-gis.org/gdata}
}
\usage{
kraje()
}
\description{
Function returning data frame of administrative units of Slovakia as \code{sf} polygons.
}
\details{
Due to stage of package development, this data is downloaded with the package.
}
