% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/okresy.R
\name{okresy}
\alias{okresy}
\title{Counties (okresy)}
\format{
\code{sf} data frame with 79 row of 2 variables + geometry:

\describe{
  \item{kraj}{Name of the region (kraj).}
  \item{okres}{Name of the county (okres).}
}
}
\source{
\url{http://www.diva-gis.org/gdata}
}
\usage{
okresy()
}
\description{
Function returning data frame of administrative units of Slovakia as \code{sf} polygons.
}
\details{
Due to stage of package development, this data is downloaded with the package.
}
