% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/republika.R
\name{republika}
\alias{republika}
\title{Borders (republika)}
\format{\code{sf} data frame with 1 row of geometry:}
\source{
\url{http://www.diva-gis.org/gdata}
}
\usage{
republika()
}
\description{
Function returning boundaries of Slovakia as \code{sf} polygon.
}
\details{
Due to stage of package development, this data is downloaded with the package.
}
\examples{
library(sf)
borders <- republika()
plot(borders, col = "blue")

}
