% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rieky.R
\name{rieky}
\alias{rieky}
\title{Rivers (rieky)}
\format{\code{sf} data frame with 396 rows of 1 variable + geometry:

\describe{
  \item{rieka}{Name of the river (rieka).}
}}
\source{
\url{http://www.diva-gis.org/gdata}
}
\usage{
rieky()
}
\description{
Function returning data frame of rivers of Slovakia as \code{sf} polygons.
}
\details{
Due to stage of package development, this data is downloaded with the package.
}
