% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zeleznice.R
\name{zeleznice}
\alias{zeleznice}
\title{Railways (zeleznice)}
\format{
\code{sf} data frame with 261 rows of 3 variables + geometry:

\describe{
  \item{rail_id}{ID of rail.}
  \item{prevadzkova}{Operation of rail.}
  \item{typ}{Type of rail.}
}
}
\source{
\url{http://www.diva-gis.org/gdata}
}
\usage{
zeleznice()
}
\description{
Function returning railways of Slovakia as \code{sf} polygon.
}
\details{
Due to stage of package development, this data is downloaded with the package.
}
